v2-connectors/v2-loldrivers/requirements.txt
pycti[connector]>=6.0.0,<7
stix2>=3.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
